Use `Long` as RowID (#87)
What changes were proposed in this pull request?

Use Long as RowID, for parquet interface support.
Delete InMemoryIndexNode and InMemoryIndexNodeValues.
Add some comments for understanding.

How was this patch tested?

FilterSuite.

* temp

* use Long for RowID, and add some comment for further code understanding

+++ src/main/scala/org/apache/spark/sql/execution/datasources/spinach/FiberCacheManager.scala
@@ -161,7 +161,7 @@ abstract class DataFile {
 
   def getFiberData(groupId: Int, fiberId: Int, conf: Configuration): FiberCacheData
   def iterator(conf: Configuration, requiredIds: Array[Int]): Iterator[InternalRow]
-  def iterator(conf: Configuration, requiredIds: Array[Int], rowIds: Array[Int])
+  def iterator(conf: Configuration, requiredIds: Array[Int], rowIds: Array[Long])
   : Iterator[InternalRow]
 }
